+++ changelog.md
@@ -6,8 +6,10 @@
 This version is not released yet and is under active development.
 ```
 
-- \[mpm\] Fix producetion of specifiers in `restore` subcommand.
-- \[mpm\] Reactivate Atom install on macOS.
+- \[mpm\] Allow managers to be flagged as deprecated.
+- \[apm\] Flag `apm` as deprecated.
+- \[mpm\] Remove Atom integration tests.
+- \[mpm\] Fix production of specifiers in `restore` subcommand.
 - \[mpm\] Fix installation of Scoop on Windows in unittests.
 - \[mpm\] Fix installation of brew on Ubuntu in unittests.
 - \[mpm\] Use form-based issue templates for bug reports and new package manager support requests.